p2pX

+++ Map/TileMap.cpp
@@ -389,11 +389,11 @@ void TileMap::render(sf::RenderTarget &target, const sf::Vector2i& gridPosition)
     }
 
 const int TileMap::getLayerSize(const int x, const int y, const int layer) const {
-    if (x >= 0 && x <= this->map.size())
+    if (x >= 0 && x <= static_cast<int>(this->map.size()))
     {
-        if (y >= 0 && this->map[x].size())
+        if (y >= 0 && static_cast<int>(this->map[x].size()))
         {
-            if (layer >= 0 && layer < this->map[x][y].size())
+            if (layer >= 0 && layer < static_cast<int>(this->map[x][y].size()))
             {
                 return this->map[x][y][layer].size();
             }
